requirements.txt
requests
python-socketio<5
python-engineio<4
pyyaml
